0xError tests (#48)
* 0x error tests

* fix test to use cowsdk

* fix tests according to comments

* better tests
